feat: 현재 상영작, 개봉 예정작 view, viewModel 구현

+++ lib/data/mapper/now_playing_mapper.dart
@@ -20,7 +20,7 @@ extension NowPlayingMapper on Results {
       title: title ?? '',
       overview: overview ?? '',
       releaseDate: releaseDate ?? '',
-      voteAverage: voteAverage?.toInt() ?? -1,
+      voteAverage: voteAverage?.toDouble() ?? -1,
     );
   }
 }
